Remove extraneous parameter from call to async_add_executor_job (#16)
* Remove extraneous parameter from call to async_add_executor_job

This removes `dpi=200` from the call to `async_add_executor_job`. Right
now this code throws an exception as `async_add_executor_job` does not have
a named parameter `dpi`. I am not sure this is fixable so long as
`async_add_executor_job` only implements `*args` and not `**kwargs`.

It should be safe to remove `dpi=200` here since the figure have `dpi=200`
specified on creation.

Fixes #14 and #15

* Use debug log messages according to coding conventions